+++ tests/any_vec.rs
@@ -2,9 +2,10 @@ use std::any::{TypeId};
 use std::mem::{size_of};
 use std::mem::forget;
 use std::ptr::NonNull;
-use itertools::assert_equal;
+use itertools::{any, assert_equal};
 use any_vec::AnyVec;
 use any_vec::any_value::{AnyValueRaw, AnyValueWrapper};
+use any_vec::mem::Stack;
 
 #[allow(dead_code)]
 unsafe fn any_as_u8_slice<T: Sized>(p: &T) -> &[u8] {
@@ -329,6 +330,27 @@ fn shrink_to_test(){
 
     any_vec.shrink_to(5);
     assert_eq!(any_vec.capacity(), 5);
+}
+
+#[test]
+fn mem_stack_test(){
+    use any_vec::traits::None;
+    type FixedAnyVec<Traits = dyn None> = AnyVec<Traits, Stack<512>>;
+
+    let mut any_vec: FixedAnyVec = AnyVec::new::<String>();
+    {
+        let mut vec = any_vec.downcast_mut::<String>().unwrap();
+        vec.push(String::from("0"));
+        vec.push(String::from("1"));
+        vec.push(String::from("2"));
+        vec.push(String::from("3"));
+    }
+
+    // Should fail to compile.
+    //any_vec.reserve(1);
+
+    assert_eq!(any_vec.capacity(), 512);
+    assert_eq!(any_vec.len(), 4);
 }
 
 #[test]